issue #1: Fix QueryableMutex to use new abilities of rubies Mutex.
As Mutexes in Ruby 1.9.x and later version has ability to detect that they was
locked by the same thread QueryableMutex can be rewritten using that
standard behaviour.

GOTCHA: It's possible to eliminate QueryableMutex class as it duplicates
behaviour of standard Mutex in Ruby 1.9.x and later versions, but it was
leaved here for now for interface compatibility purpose.
Later in major version bump it can be removed from "needle".